Create DeviceCodeRegistry.cpp and place into cudaq-common.so
